requirements.txt
awscli==1.16.226
botocore==1.12.216
colorama==0.3.9
docutils==0.14
futures==3.2.0
http==0.2
jmespath==0.9.3
pyasn1==0.4.4
python-dateutil==2.7.3
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.1
six==1.11.0
urllib3==1.23
virtualenv==16.7.7
